requirements.txt
djangorestframework==3.15.2
Django==4.2.18
psycopg2-binary==2.9.10
whitenoise==6.7.0
gunicorn==23.0.0
firebase-admin==6.6.0
google-auth==2.38.0
sentence-transformers==3.4.0
torch==2.5.1
scikit-learn==1.6.1 
PyMuPDF==1.24.5